\docType{data}
\name{Bali}
\alias{Bali}
\title{Jemaah Islamiyah 2002 Bali Terrorist Network}
\format{A network object with 17 vertices and 126 edges}
\description{
  This network shows the interactions among the Jemaah
  Islamiyah terrorist group that carried out the bombings
  in Bali in 2002.
}
\details{
  To be added later. [Comment added 2015]
}
\section{Node characteristics.}{
  \itemize{ \item \code{vertex.names} - Last name \item
  \code{role} - Role in group: \itemize{ \item \emph{CT} -
  Command Team \item \emph{OA} - Operational Assistant
  \item \emph{BM} - Bomb Maker \item \emph{SB} - Suicide
  Bomber \item \emph{TL} - Team Lima } }
}

\section{Edge characteristics.}{
  \itemize{ \item \code{IC} - Interactional criteria (1 to
  5) }
}
\author{
  Douglas Luke
}
\references{
  \url{http:eprints.qut.edu.au/16591}
}
\keyword{datasets}

