\docType{data}
\name{DHHS}
\alias{DHHS}
\title{DHHS Collaboration Network}
\format{A network object with 54 nodes and 477 edges}
\description{
  This data set contains the DHHS Collaboration network.
  This is an undirected network with 54 nodes and 477
  edges.
}
\details{
  Contains the following vertex attributes: \itemize{ \item
  agency. [X] categories \item dum. Add info }
}
\author{
  Douglas Luke
}
\references{
  Leischow, S. J., Luke, D.A., et al. (2010). Mapping U.S. government 
  tobacco control leadership: Networked for success? 
  \emph{Nicotine & Tobacco Research, 12, 888-894}.
}

\keyword{datasets}

