\docType{data}
\name{Krebs}
\alias{Krebs}
\title{9-11 hijackers network}
\format{A network object with 19 vertices and 27 edges}
\description{
  This network shows the contacts among the terrorists who
  carried out the 9-11 hijackings in 2001.
}
\details{
  This network data set was first constructed and analyzed
  by Valdis Krebs. Using publically available data at the time,
  he constructed this contact network to illustrate the contact ties
  that the 19 hijackers had prior to the attack.
}
\section{Node characteristics:}{
  \itemize{ \item \code{name} - Full name \item
  \code{flight} - Flight number }
}


\references{
Krebs, Valdis. Uncloaking Terrorist Networks. First Monday, [S.l.], apr. 2002. ISSN 13960466. Available at: \url{http://firstmonday.org/ojs/index.php/fm/article/view/941/863}.
  }
\keyword{datasets}

