\docType{data}
\name{middleschool}
\alias{middleschool}
\title{Middleschool Friendship Network}
\format{A network object with 37 vertices and 179 arcs}
\description{
  This network object contains friendship ties among a
  class of 6th grade students in California.
}
\details{
  These data are described in Thomas Valente's book
  \emph{Social Networks and Health}. See Figure 1-1 on page
  4. The two non-numbered isolates in this figure were
  dropped from the dataset. Students were asked to nominate
  their five closest friends in their class. Thus, this is
  a directed network.
}

\section{Node characteristics:}{
  \itemize{ \item \code{gender} - Student's gender (1=female,
  2=male)
  }
}

\references{
  Valente, T. W. (2010). \emph{Social Networks and Health:
  Models, Methods, and Applications.} Oxford University
  Press.
}
\keyword{datasets}

