\docType{data}
\name{Moreno}
\alias{Moreno}
\title{Moreno Sociogram}
\format{A network object with 33 vertices and 46 edges}
\description{
  This network object contains friendship ties among a
  4th grade class.
}
\details{
  These data are contained in a sociogram constructed by
  Jacob Moreno, and published in the New York Times in
  1933. The original diagram suggests that the graph
  is directed, but the network object here is non-directed.
}
\section{Node characteristics:}{
  \itemize{ \item \code{gender} - Student's gender (1=female,
  2=male)
  }
}
\references{
  Moreno, J. L. (1934). \emph{Who shall survive? A new approach to the problem of
  human interrelations.} Nervous and mental disease monograph series,
  no. 58. Washington, DC: Nervous and Mental Disease Publishing Co.

}
\keyword{datasets}

