\docType{data}
\name{PowerGrid}
\alias{PowerGrid}
\title{Power Grid boardgame network}
\format{A network object with 42 vertices and 87 edges}
\description{
  This network object contains the connections among U.S.
  cities as they appear in the boardgame Power Grid,
  published by Rio Grande games.
}
\section{Node characteristics}{
  \itemize{ \item \code{vertex.name} - Name of the city.  }
}

\section{Node characteristics:}{
  \itemize{ \item \code{cost} - Cost to connect the two
  cities, according to the game rules.  }
}
\author{
  Douglas Luke
}
\references{
  \url{http://boardgamegeek.com/boardgame/2651/power-grid}
}
\keyword{datasets}

