\docType{data}
\name{Bali}
\alias{Bali}
\title{Jemaah Islamiyah 2002 Bali Terrorist Network}
\format{A network object with 17 vertices and 126 edges}
\description{
  This network shows the interactions among the Jemaah
  Islamiyah terrorist group that carried out the bombings
  in Bali in 2002.
}
\details{
  These data were originally collected and created by Stuart
  Koschade as part of his dissertation on the social network
  analysis of terrorist groups. The ties here represent
  contacts among the Bali terrorist cell. The dataset
  also includes an edge characteristic, \emph{IC},
  that measures the frequency and duration of the contact,
  where 1 indicates a weak relationship, and 5 indicates
  a strong relationship.
}
\section{Node characteristics:}{
  \itemize{ \item \code{vertex.names} - Last name \item
  \code{role} - Role in group: \itemize{ \item \emph{CT} -
  Command Team \item \emph{OA} - Operational Assistant
  \item \emph{BM} - Bomb Maker \item \emph{SB} - Suicide
  Bomber \item \emph{TL} - Team Lima } }
}

\section{Edge characteristics:}{
  \itemize{ \item \code{IC} - Interactional criteria (1 to
  5) }
}

\references{
  Koschade, S. A. (2007). \emph{The internal dynamics of terrorist cells:
  a social network analysis of terrorist cells in an Australian context.}
  PhD thesis, Queensland University of Technology. \url{http:eprints.qut.edu.au/16591}
}
\keyword{datasets}

