\name{ICTS_G10}
\alias{ICTS_G10}
\docType{data}
\title{ICTS Grants Collaboration}


\format{A network object with 493 vertices and 1359 edges}

\description{
  This network object contains collaboration ties among 493 members of
  Washington University's Institute of Clinical and Translational
  Sciences (ICTS).
}

\details{
These data were collected as part of the evaluation of the ICTS.
A tie exists between two nodes if those scientists worked
  together on a grant submission in 2010.
}

\section{Node characteristics:}{
  \itemize{ \item \code{vertex.names} - ICTS member ID
  \item \code{year} - Year member joined ICTS
  \item \code{DiscCol} - Discipline code, collapsed
  \item \code{DiscExp} - Discipline code, expanded

  }
}

\author{
  Douglas Luke
}

\references{
  Luke, D.A., et al. (2014). Breaking down silos:
  Mapping growth of cross-cisciplinary collaboration in a translational science initiative
  \emph{Clinical and Translational Science, 8,} 143-149.

}
\keyword{datasets}
